+++ agent/api/v1/health.go
@@ -4,15 +4,21 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/erikmagkekse/btrfs-nfs-csi/agent/storage"
 	"github.com/labstack/echo/v5"
 )
 
-func Healthz(version, commit string, features map[string]string) echo.HandlerFunc {
+func Healthz(version, commit string, features map[string]string, store *storage.Storage) echo.HandlerFunc {
 	startTime := time.Now()
 
 	return func(c *echo.Context) error {
+		status := HealthStatusOK
+		if store.IsDegraded() {
+			status = HealthStatusDegraded
+		}
+
 		return c.JSON(http.StatusOK, HealthResponse{
-			Status:        "ok", // Maybe for the future we can add more status levels like "degraded"
+			Status:        status,
 			Version:       version,
 			Commit:        commit,
 			UptimeSeconds: int(time.Since(startTime).Seconds()),